Split out user controller preference methods to new controller
